Move program specific JSONForms controls out of the common pkg
