+++ client/packages/common/src/ui/forms/JsonForms/components/index.ts
@@ -4,9 +4,6 @@ export * from './Label';
 export * from './Text';
 export * from './Select';
 export * from './Date';
-export * from './DateOfBirth';
 export * from './Array';
 export * from './FirstItemArray';
 export * from './Boolean';
-export * from './IdGenerator';
-export * from './EncounterLineChart';
