+++ client/packages/system/src/Patient/JsonForms/api/hooks/document/useEncounterFields.ts
