+++ client/packages/system/src/Patient/ProgramEnrolment/DetailModal.tsx
@@ -2,14 +2,13 @@ import React, { FC } from 'react';
 import {
   BasicSpinner,
   DialogButton,
-  SaveDocumentMutation,
   useDialog,
-  useJsonForms,
 } from '@openmsupply-client/common';
 import { useProgramEnrolment } from './api/hooks';
 import { usePatientModalStore } from '../hooks';
 import { PatientModal } from '../PatientView';
 import { usePatient } from '../api';
+import { SaveDocumentMutation, useJsonForms } from '../JsonForms';
 
 const useUpsertProgramEnrolment = (
   patientId: string,
